# eskipfile: optimize Watch.LoadUpdates