Do not parse routes if file content did not change since last successful read.

Note that after https://github.com/zalando/skipper/pull/2304 RemoteWatch does not update file if response Etag did not change but underlying Watch would still parse routes.

```
goos: linux
goarch: amd64
pkg: github.com/zalando/skipper/eskipfile
cpu: Intel(R) Core(TM) i5-8350U CPU @ 1.70GHz
                  │     HEAD~1     │                 HEAD                 │
                  │     sec/op     │    sec/op     vs base                │
WatchLoadUpdate-8   17889.7µ ± 24%   220.0µ ± 24%  -98.77% (p=0.000 n=10)

                  │    HEAD~1     │                 HEAD                 │
                  │     B/op      │     B/op      vs base                │
WatchLoadUpdate-8   7837.4Ki ± 0%   528.5Ki ± 0%  -93.26% (p=0.000 n=10)

                  │     HEAD~1      │                HEAD                │
                  │    allocs/op    │ allocs/op   vs base                │
WatchLoadUpdate-8   110110.000 ± 0%   6.000 ± 0%  -99.99% (p=0.000 n=10)
```